Update from Cinder

+++ cinder/cmd/manage.py
@@ -58,8 +58,10 @@ from __future__ import print_function
 import logging as python_logging
 import os
 import sys
+import time
 
 from oslo_config import cfg
+from oslo_db import exception as db_exc
 from oslo_db.sqlalchemy import migration
 from oslo_log import log as logging
 import oslo_messaging as messaging
@@ -70,6 +72,7 @@ i18n.enable_lazy()
 
 # Need to register global_opts
 from cinder.common import config  # noqa
+from cinder.common import constants
 from cinder import context
 from cinder import db
 from cinder.db import migration as db_migration
@@ -78,7 +81,6 @@ from cinder import exception
 from cinder.i18n import _
 from cinder import objects
 from cinder import rpc
-from cinder import utils
 from cinder import version
 from cinder.volume import utils as vutils
 
@@ -116,7 +118,7 @@ class ShellCommands(object):
         """
         self.run('python')
 
-    @args('--shell', dest="shell",
+    @args('--shell',
           metavar='<bpython|ipython|python>',
           help='Python shell')
     def run(self, shell=None):
@@ -205,6 +207,9 @@ class HostCommands(object):
 class DbCommands(object):
     """Class for managing the database."""
 
+    online_migrations = (db.migrate_consistencygroups_to_groups,
+                         db.migrate_add_message_prefix)
+
     def __init__(self):
         pass
 
@@ -228,8 +233,66 @@ class DbCommands(object):
         if age_in_days <= 0:
             print(_("Must supply a positive, non-zero value for age"))
             sys.exit(1)
+        if age_in_days >= (int(time.time()) / 86400):
+            print(_("Maximum age is count of days since epoch."))
+            sys.exit(1)
         ctxt = context.get_admin_context()
-        db.purge_deleted_rows(ctxt, age_in_days)
+
+        try:
+            db.purge_deleted_rows(ctxt, age_in_days)
+        except db_exc.DBReferenceError:
+            print(_("Purge command failed, check cinder-manage "
+                    "logs for more details."))
+            sys.exit(1)
+
+    def _run_migration(self, ctxt, max_count, ignore_state):
+        ran = 0
+        for migration_meth in self.online_migrations:
+            count = max_count - ran
+            try:
+                found, done = migration_meth(ctxt, count, ignore_state)
+            except Exception:
+                print(_("Error attempting to run %(method)s") %
+                      {'method': migration_meth.__name__})
+                found = done = 0
+
+            if found:
+                print(_('%(total)i rows matched query %(meth)s, %(done)i '
+                        'migrated') % {'total': found,
+                                       'meth': migration_meth.__name__,
+                                       'done': done})
+            if max_count is not None:
+                ran += done
+                if ran >= max_count:
+                    break
+        return ran
+
+    @args('--max_count', metavar='<number>', dest='max_count', type=int,
+          help='Maximum number of objects to consider.')
+    @args('--ignore_state', action='store_true', dest='ignore_state',
+          help='Force records to migrate even if another operation is '
+               'performed on them. This may be dangerous, please refer to '
+               'release notes for more information.')
+    def online_data_migrations(self, max_count=None, ignore_state=False):
+        """Perform online data migrations for the release in batches."""
+        ctxt = context.get_admin_context()
+        if max_count is not None:
+            unlimited = False
+            if max_count < 1:
+                print(_('Must supply a positive value for max_number.'))
+                sys.exit(127)
+        else:
+            unlimited = True
+            max_count = 50
+            print(_('Running batches of %i until complete.') % max_count)
+
+        ran = None
+        while ran is None or ran != 0:
+            ran = self._run_migration(ctxt, max_count, ignore_state)
+            if not unlimited:
+                break
+
+        sys.exit(1 if ran else 0)
 
 
 class VersionCommands(object):
@@ -255,7 +318,7 @@ class VolumeCommands(object):
         if self._client is None:
             if not rpc.initialized():
                 rpc.init(CONF)
-                target = messaging.Target(topic=CONF.volume_topic)
+                target = messaging.Target(topic=constants.VOLUME_TOPIC)
                 serializer = objects.base.CinderObjectSerializer()
                 self._client = rpc.get_client(target, serializer=serializer)
 
@@ -429,13 +492,23 @@ class BackupCommands(object):
             bk.save()
 
 
-class ServiceCommands(object):
+class BaseCommand(object):
+    @staticmethod
+    def _normalize_time(time_field):
+        return time_field and timeutils.normalize_time(time_field)
+
+    @staticmethod
+    def _state_repr(is_up):
+        return ':-)' if is_up else 'XXX'
+
+
+class ServiceCommands(BaseCommand):
     """Methods for managing services."""
     def list(self):
         """Show a list of all cinder services."""
         ctxt = context.get_admin_context()
         services = objects.ServiceList.get_all(ctxt)
-        print_format = "%-16s %-36s %-16s %-10s %-5s %-20s %-12s %-15s"
+        print_format = "%-16s %-36s %-16s %-10s %-5s %-20s %-12s %-15s %-36s"
         print(print_format % (_('Binary'),
                               _('Host'),
                               _('Zone'),
@@ -443,23 +516,19 @@ class ServiceCommands(object):
                               _('State'),
                               _('Updated At'),
                               _('RPC Version'),
-                              _('Object Version')))
+                              _('Object Version'),
+                              _('Cluster')))
         for svc in services:
-            alive = utils.service_is_up(svc)
-            art = ":-)" if alive else "XXX"
-            status = 'enabled'
-            if svc.disabled:
-                status = 'disabled'
-            updated_at = svc.updated_at
-            if updated_at:
-                updated_at = timeutils.normalize_time(updated_at)
-            rpc_version = (svc.rpc_current_version or
-                           rpc.LIBERTY_RPC_VERSIONS.get(svc.binary, ''))
-            object_version = (svc.object_current_version or 'liberty')
+            art = self._state_repr(svc.is_up)
+            status = 'disabled' if svc.disabled else 'enabled'
+            updated_at = self._normalize_time(svc.updated_at)
+            rpc_version = svc.rpc_current_version
+            object_version = svc.object_current_version
+            cluster = svc.cluster_name or ''
             print(print_format % (svc.binary, svc.host.partition('.')[0],
                                   svc.availability_zone, status, art,
-                                  updated_at, rpc_version,
-                                  object_version))
+                                  updated_at, rpc_version, object_version,
+                                  cluster))
 
     @args('binary', type=str,
           help='Service to delete from the host.')
@@ -475,14 +544,137 @@ class ServiceCommands(object):
             print(_("Host not found. Failed to remove %(service)s"
                     " on %(host)s.") %
                   {'service': binary, 'host': host_name})
-            print (u"%s" % e.args)
+            print(u"%s" % e.args)
             return 2
         print(_("Service %(service)s on host %(host)s removed.") %
               {'service': binary, 'host': host_name})
 
+
+class ClusterCommands(BaseCommand):
+    """Methods for managing clusters."""
+    def list(self):
+        """Show a list of all cinder services."""
+        ctxt = context.get_admin_context()
+        clusters = objects.ClusterList.get_all(ctxt, services_summary=True)
+        print_format = "%-36s %-16s %-10s %-5s %-20s %-7s %-12s %-20s"
+        print(print_format % (_('Name'),
+                              _('Binary'),
+                              _('Status'),
+                              _('State'),
+                              _('Heartbeat'),
+                              _('Hosts'),
+                              _('Down Hosts'),
+                              _('Updated At')))
+        for cluster in clusters:
+            art = self._state_repr(cluster.is_up)
+            status = 'disabled' if cluster.disabled else 'enabled'
+            heartbeat = self._normalize_time(cluster.last_heartbeat)
+            updated_at = self._normalize_time(cluster.updated_at)
+            print(print_format % (cluster.name, cluster.binary, status, art,
+                                  heartbeat, cluster.num_hosts,
+                                  cluster.num_down_hosts, updated_at))
+
+    @args('--recursive', action='store_true', default=False,
+          help='Delete associated hosts.')
+    @args('binary', type=str,
+          help='Service to delete from the cluster.')
+    @args('cluster-name', type=str, help='Cluster to delete.')
+    def remove(self, recursive, binary, cluster_name):
+        """Completely removes a cluster."""
+        ctxt = context.get_admin_context()
+        try:
+            cluster = objects.Cluster.get_by_id(ctxt, None, name=cluster_name,
+                                                binary=binary,
+                                                get_services=recursive)
+        except exception.ClusterNotFound:
+            print(_("Couldn't remove cluster %s because it doesn't exist.") %
+                  cluster_name)
+            return 2
+
+        if recursive:
+            for service in cluster.services:
+                service.destroy()
+
+        try:
+            cluster.destroy()
+        except exception.ClusterHasHosts:
+            print(_("Couldn't remove cluster %s because it still has hosts.") %
+                  cluster_name)
+            return 2
+
+        msg = _('Cluster %s successfully removed.') % cluster_name
+        if recursive:
+            msg = (_('%(msg)s And %(num)s services from the cluster were also '
+                     'removed.') % {'msg': msg, 'num': len(cluster.services)})
+        print(msg)
+
+    @args('--full-rename', dest='partial',
+          action='store_false', default=True,
+          help='Do full cluster rename instead of just replacing provided '
+               'current cluster name and preserving backend and/or pool info.')
+    @args('current', help='Current cluster name.')
+    @args('new', help='New cluster name.')
+    def rename(self, partial, current, new):
+        """Rename cluster name for Volumes and Consistency Groups.
+
+        Useful when you want to rename a cluster, particularly when the
+        backend_name has been modified in a multi-backend config or we have
+        moved from a single backend to multi-backend.
+        """
+        ctxt = context.get_admin_context()
+
+        # Convert empty strings to None
+        current = current or None
+        new = new or None
+
+        # Update Volumes
+        num_vols = objects.VolumeList.include_in_cluster(
+            ctxt, new, partial_rename=partial, cluster_name=current)
+
+        # Update Consistency Groups
+        num_cgs = objects.ConsistencyGroupList.include_in_cluster(
+            ctxt, new, partial_rename=partial, cluster_name=current)
+
+        if num_vols or num_cgs:
+            msg = _('Successfully renamed %(num_vols)s volumes and '
+                    '%(num_cgs)s consistency groups from cluster %(current)s '
+                    'to %(new)s')
+            print(msg % {'num_vols': num_vols, 'num_cgs': num_cgs, 'new': new,
+                         'current': current})
+        else:
+            msg = _('No volumes or consistency groups exist in cluster '
+                    '%(current)s.')
+            print(msg % {'current': current})
+            return 2
+
+
+class ConsistencyGroupCommands(object):
+    """Methods for managing consistency groups."""
+
+    @args('--currenthost', required=True, help='Existing CG host name')
+    @args('--newhost', required=True, help='New CG host name')
+    def update_cg_host(self, currenthost, newhost):
+        """Modify the host name associated with a Consistency Group.
+
+        Particularly to recover from cases where one has moved
+        a host from single backend to multi-backend, or changed the host
+        configuration option, or modified the backend_name in a multi-backend
+        config.
+        """
+
+        ctxt = context.get_admin_context()
+        groups = objects.ConsistencyGroupList.get_all(
+            ctxt, {'host': currenthost})
+        for gr in groups:
+            gr.host = newhost
+            gr.save()
+
+
 CATEGORIES = {
     'backup': BackupCommands,
     'config': ConfigCommands,
+    'cluster': ClusterCommands,
+    'cg': ConsistencyGroupCommands,
     'db': DbCommands,
     'host': HostCommands,
     'logs': GetLogCommands,
@@ -532,29 +724,32 @@ category_opt = cfg.SubCommandOpt('category',
 
 
 def get_arg_string(args):
-    arg = None
     if args[0] == '-':
         # (Note)zhiteng: args starts with FLAGS.oparser.prefix_chars
         # is optional args. Notice that cfg module takes care of
         # actual ArgParser so prefix_chars is always '-'.
         if args[1] == '-':
             # This is long optional arg
-            arg = args[2:]
+            args = args[2:]
         else:
-            arg = args[1:]
-    else:
-        arg = args
+            args = args[1:]
 
-    return arg
+    # We convert dashes to underscores so we can have cleaner optional arg
+    # names
+    if args:
+        args = args.replace('-', '_')
+
+    return args
 
 
 def fetch_func_args(func):
-    fn_args = []
+    fn_kwargs = {}
     for args, kwargs in getattr(func, 'args', []):
-        arg = get_arg_string(args[0])
-        fn_args.append(getattr(CONF.category, arg))
+        # Argparser `dest` configuration option takes precedence for the name
+        arg = kwargs.get('dest') or get_arg_string(args[0])
+        fn_kwargs[arg] = getattr(CONF.category, arg)
 
-    return fn_args
+    return fn_kwargs
 
 
 def main():
@@ -579,19 +774,11 @@ def main():
     except cfg.ConfigDirNotFoundError as details:
         print(_("Invalid directory: %s") % details)
         sys.exit(2)
-    except cfg.ConfigFilesNotFoundError:
-        cfgfile = CONF.config_file[-1] if CONF.config_file else None
-        if cfgfile and not os.access(cfgfile, os.R_OK):
-            st = os.stat(cfgfile)
-            print(_("Could not read %s. Re-running with sudo") % cfgfile)
-            try:
-                os.execvp('sudo', ['sudo', '-u', '#%s' % st.st_uid] + sys.argv)
-            except Exception:
-                print(_('sudo failed, continuing as if nothing happened'))
-
-        print(_('Please re-run cinder-manage as root.'))
+    except cfg.ConfigFilesNotFoundError as e:
+        cfg_files = e.config_files
+        print(_("Failed to read configuration file(s): %s") % cfg_files)
         sys.exit(2)
 
     fn = CONF.category.action_fn
-    fn_args = fetch_func_args(fn)
-    fn(*fn_args)
+    fn_kwargs = fetch_func_args(fn)
+    fn(**fn_kwargs)
